Fixed AllocateAndSpreadRequest problem
We used to use the same AllocationRequest message both in fileTransferActor and durinc create/import. This was a bad thing, since during create/import a spread of the tags was in order, while this must not be done in the fileTransferActor, otherwise the owner of the file changes before the file itself has been transferred, and this may lead to error if an error occur during the transfer.